services/task_orchestrator_service/requirements.txt
# Core Dependencies
pydantic>=2.6.1
python-dotenv>=1.0.0
aio-pika>=9.3.0  # Async RabbitMQ client
structlog>=24.1.0  # Structured logging
tenacity>=8.2.3  # For retry mechanisms
prometheus-client>=0.19.0  # For metrics
opentelemetry-api>=1.22.0  # For tracing
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-aio-pika>=0.42b0  # RabbitMQ instrumentation

# Database
asyncpg>=0.29.0  # Async PostgreSQL driver
sqlalchemy[asyncio]>=2.0.27
alembic>=1.13.1  # For database migrations

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0  # For test coverage
pytest-mock>=3.12.0  # For mocking in tests
pytest-env>=1.1.3  # For environment variables in tests
pytest-timeout>=2.2.0  # For test timeouts

# Development
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0  # For git hooks 